requirements.txt
aioredis==1.3.0
amqp==2.5.1
asgiref==3.2.2
asn1crypto==1.0.1
async-timeout==3.0.1
attrs==19.2.0
autobahn==19.10.1
Automat==0.7.0
Babel==2.7.0
billiard==3.6.1.0
boto3==1.10.23
botocore==1.13.50
celery==4.3.0
certifi==2019.6.16
cffi==1.12.3
channels==2.3.0
channels-redis==2.4.0
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
daphne==2.3.0
Django==2.2.3
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-cors-headers==3.0.2
django-filter==2.2.0
django-jinja==2.4.1
django-rest-passwordreset==1.1.0rc3
django-rest-swagger==2.1.0
django-ses==1.0.3
django-storages==1.7.2
django-timezone-field==3.0
django-uuidfield==0.5.0
djangorestframework==3.10.1
djangorestframework-gis==0.14
djangorestframework-simplejwt==4.3.0
docutils==0.15.2
fcm-django==0.2.21
flower==0.9.3
future==0.18.2
gevent==1.4.0
greenlet==0.4.15
hiredis==1.0.0
humanize==0.5.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.19
incremental==17.5.0
itypes==1.1.0
Jinja2==2.10.1
jmespath==0.10.0
kombu==4.6.4
Markdown==3.1.1
MarkupSafe==1.1.1
msgpack==0.6.2
openapi-codec==1.3.2
Pillow==6.1.0
plumbum==1.6.7
psycopg2==2.8.3
pycparser==2.19
pyfcm==1.4.5
PyHamcrest==1.9.0
PyJWT==1.7.1
PySocks==1.7.0
python-crontab==2.3.8
python-dateutil==2.8.0
pytz==2019.1
redis==3.3.8
requests==2.22.0
requests-toolbelt==0.9.1
rpyc==4.1.1
s3transfer==0.2.1
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
tornado==5.1.1
twilio==6.29.2
Twisted==20.3.0
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
zipp==0.5.2
zope.interface==4.6.0
